separated theme and app's root
