+++ app/android/build.gradle.kts
@@ -20,7 +20,7 @@ kotlin {
             implementation(libs.androidx.activityCompose)
             implementation(libs.kotlinx.coroutines.android)
             implementation(libs.ktor.client.okhttp)
-            implementation(project(":composeApp"))
+            implementation(project(":presentation:root"))
         }
     }
 }
