+++ app/desktop/build.gradle.kts
@@ -17,7 +17,7 @@ kotlin {
             implementation(compose.desktop.currentOs)
             implementation(libs.kotlinx.coroutines.swing)
             implementation(libs.ktor.client.okhttp)
-            implementation(project(":composeApp"))
+            implementation(project(":presentation:root"))
         }
     }
 }
