+++ settings.gradle.kts
@@ -33,7 +33,8 @@ plugins {
     id("org.gradle.toolchains.foojay-resolver-convention").version("0.10.0")
 }
 
-include(":composeApp")
 include(":app:android")
 include(":app:desktop")
 include(":app:wasmJs")
+include(":presentation:theme")
+include(":presentation:root")
